20240920-hs_community

+++ worldSnackDiscountStore/src/main/java/com/worldsnack/dto/CommentDTO.java
@@ -6,10 +6,10 @@ import lombok.Data;
 
 @Data
 public class CommentDTO {
-	private Long comment_idx;					// 댓글 인덱스
-	private Long post_id;							// 게시글 인덱스
-  private Long user_idx;						// 작성자 인덱스
-  private Long parent_comment_idx;  // 부모 댓글 인덱스, null 가능
+	private int comment_idx;					// 댓글 인덱스
+	private int post_id;							// 게시글 인덱스
+  private int user_idx;						// 작성자 인덱스
+  private Integer parent_comment_idx;  // 부모 댓글 인덱스, null 가능
   private String comment_text;			// 댓글 내용
   private Date comment_date;				// 작성된 시간
   private Date comment_update_date; // 수정된 시간, null 가능
@@ -23,8 +23,8 @@ public class CommentDTO {
   private int reported_count;       // 댓글이 신고된 수
   
   private Integer vote_type; 				// 현재 사용자가 한 투표 타입 ('up' or 'down')
-  private Long vote_idx;    				// 투표 인덱스
+  private int vote_idx;    				// 투표 인덱스
   
   private boolean isAuthor;
-  private Long currentUserId;
+  private int currentUserId;
 }
